refactor: crete auth in utils for registerWithEmailAndPassword
