+++ graphiql-app/src/components/SignUpForm/SignUpForm.tsx
@@ -2,40 +2,19 @@
 
 import { Alert, Box, Button, TextField, Typography } from '@mui/material';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
 import { useAuth } from '@/hooks/useAuth';
 import { ISignUpFormData } from '@/types/formsType';
-import { auth } from '@/utils/fireBaseConfig';
+import { registerWithEmailAndPassword } from '@/utils/auth';
 import { schemaSignUp } from '@/utils/validationSchema';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { createUserWithEmailAndPassword, updateProfile } from 'firebase/auth';
 import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/navigation';
 
 import { SERVICE_MESSAGES } from '../../constants/SERVICE_MESSAGES';
 import PasswordStrength from '../passwordStrength/passwordStrength';
-
-const registerWithEmailAndPassword = async (
-  { name, email, password }: ISignUpFormData,
-  setError: React.Dispatch<React.SetStateAction<string | null>>,
-) => {
-  try {
-    const res = await createUserWithEmailAndPassword(auth, email, password);
-
-    if (auth.currentUser) {
-      await updateProfile(auth.currentUser, { displayName: name });
-      return true;
-    }
-
-    document.cookie = `userid=${res.user.uid}`;
-  } catch (err) {
-    if (err instanceof Error) {
-      setError(err.message);
-    }
-  }
-};
 
 const SignUpForm = () => {
   const [error, setError] = useState<string | null>(null);
@@ -48,9 +27,12 @@ const SignUpForm = () => {
     resolver: yupResolver(schemaSignUp),
     mode: 'onChange',
   });
+
   const router = useRouter();
   const { user } = useAuth();
   const password = watch('password');
+  const t = useTranslations('signUp');
+
   const onSubmit: SubmitHandler<ISignUpFormData> = async data => {
     setError(null);
     const { name, email, password } = data;
@@ -62,11 +44,11 @@ const SignUpForm = () => {
     if (userCredential === true) router.push('/');
   };
 
-  const t = useTranslations('signUp');
-
-  if (user) {
-    router.push('/');
-  }
+  useEffect(() => {
+    if (user) {
+      router.push('/');
+    }
+  }, [router, user]);
 
   return (
     <Box
